requirements.txt
langchain
python-dotenv
openai
google-generativeai
langchain-openai
streamlit